Создан сервис для аутентификации пользователей Обновлён (дополнен) ProductController

+++ src/main/java/com/example/springproject/controller/ProductController.java
@@ -3,6 +3,8 @@ package com.example.springproject.controller;
 
 import com.example.springproject.entity.Product;
 import com.example.springproject.service.ProductService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,23 +19,50 @@ public class ProductController {
     }
 
     @GetMapping
-    List<Product> getAllProducts() {
-        return productService.getAllProducts();
+    public ResponseEntity<List<Product>> getAllProducts() {
+        List<Product> products = productService.getAllProducts();
+        return ResponseEntity.ok(products);
     }
 
     @GetMapping("/{id}")
-    public Product getProductById(@PathVariable Long id) {
-        return productService.getProductById(id);
+    public ResponseEntity<Product> getProductById(@PathVariable Long id) {
+        Product product = productService.getProductById(id);
+        if (product == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(product);
+    }
+
+    @GetMapping("/seller/{id}")
+    public ResponseEntity<List<Product>> getProductBySellerId(@PathVariable Long id) {
+        List<Product> products = productService.getProductBySeller(id);
+        return ResponseEntity.ok(products);
     }
 
     @PostMapping
-    public Product saveProduct(Product product) {
-        return productService.saveProduct(product);
+    public ResponseEntity<Product> saveProduct(Product product) {
+        Product savedProduct = productService.saveProduct(product);
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedProduct);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody Product product) {
+        try {
+            Product updatedProduct = productService.updateProduct(product, id);
+            return ResponseEntity.ok(updatedProduct);
+        } catch (Exception e) {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteProduct(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+        Product product = productService.getProductById(id);
+        if (product == null) {
+            return ResponseEntity.notFound().build();
+        }
         productService.deleteProduct(id);
+        return ResponseEntity.noContent().build();
     }
 
 }
